Basic soldier table UI using Tkinter
There are two working versions of the UI. The first uses basic Tkinter Entries to form a table. The second uses the dmnfarrell/tkintertable library to create a better looking, albeit less flexible, table. I'm sticking with the second version for now as I'm not well-versed in Tkinter, but that will likely change in the future.